+++ app/controllers/users/confirmations_controller.rb
@@ -2,5 +2,6 @@
 
 module Users
   class ConfirmationsController < Devise::ConfirmationsController
+    def check; end
   end
 end