+++ app/controllers/users/registrations_controller.rb
@@ -4,8 +4,8 @@ module Users
   class RegistrationsController < Devise::RegistrationsController
     protected
 
-    def after_sign_up_path_for(_resource)
-      root_path
+    def after_inactive_sign_up_path_for(resource)
+      check_confirmations_path(email: resource.email)
     end
   end
 end