+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - frontend translating with ttag
 - frontend validation for portfolio name
+- confirmation email page
 
 ## [1.0.2] - 19.01.2021
 ### Added